added closing and opening currency values

+++ src/main/java/dev/ElifnazArc/MT940_to_Excel/controller/TransactionController.java
@@ -2,11 +2,14 @@ package dev.ElifnazArc.MT940_to_Excel.controller;
 
 import dev.ElifnazArc.MT940_to_Excel.entity.Transaction;
 import dev.ElifnazArc.MT940_to_Excel.repository.TransactionRepository;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.Optional;
 
+@CrossOrigin(origins = "*")
 @RestController
 @RequestMapping("/api/transactions")
 public class TransactionController {
@@ -17,44 +20,30 @@ public class TransactionController {
     }
 
     @GetMapping
-    List<Transaction> findAll() {
+    public List<Transaction> findAll() {
         return transactionRepository.findAll();
     }
 
     @GetMapping("/{id}")
-    Transaction findById(@PathVariable int id) {
+    public ResponseEntity<Transaction> findById(@PathVariable int id) {
         Optional<Transaction> transaction = transactionRepository.findById(id);
-        return transaction.orElse(null);
+        return transaction.map(ResponseEntity::ok)
+                .orElseGet(() -> ResponseEntity.status(HttpStatus.NOT_FOUND).body(null));
     }
 
     @PostMapping
-    void save(@RequestBody Transaction transaction) {
-        transactionRepository.save(transaction);
+    public ResponseEntity<Transaction> save(@RequestBody Transaction transaction) {
+        Transaction savedTransaction = transactionRepository.save(transaction);
+        return ResponseEntity.status(HttpStatus.CREATED).body(savedTransaction);
     }
 
     @DeleteMapping("/{id}")
-    void delete(@PathVariable int id) {
-        transactionRepository.deleteById(id);
+    public ResponseEntity<Void> delete(@PathVariable int id) {
+        if (transactionRepository.existsById(id)) {
+            transactionRepository.deleteById(id);
+            return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+        } else {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
     }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 }
